Crate effect_monad

This module contains Purescript-inspired effects monads for rust

Here, an effect is defined an evaluatable function.

Macros

effect_map!

Structs

BoundEffect
A struct representing two bound effects. Ideally, we would be able to a closure here, but that's not possible without returning a boxed version of the closure, which we don't want to do.

Enums

ResolveFn
Helper enum for acting as a resolve function.

Traits

EffectMonad
Monad trait for effect functions