Enum effect_monad::ResolveFn

pub enum ResolveFn<T> {
    Const(T),
}

Helper enum for acting as a resolve function.

Ideally, we would use a closure instead of this type, but this type exists as a workaround alternative to avoid using boxed closures

Variants

Const