Struct effect_monad::BoundEffect

pub struct BoundEffect<Ea, F> {
    // some fields omitted
}

A struct representing two bound effects. Ideally, we would be able to a closure here, but that's not possible without returning a boxed version of the closure, which we don't want to do.